Improve performance of intersection detection
Intersection detection is one of the bigger performance issues,
because the algorithm looks for an intersection between each fertile
worm.
We can do some preliminary filtering of worms that aren't close to each
other. If one worm isn't close to the other one on the X or Y axis, then
there is no point in looking for intersection. Thus we can filter on
proximity.

Also, do some trivial refactoring of `World`.